Change playback row from int to long
Might be needed for songs with one pattern and many ticks per second.

+++ include/camoto/gamemusic/playback.hpp
@@ -51,7 +51,7 @@ class Playback: virtual public SynthPCMCallback
 			unsigned int order;
 
 			/// Row index within pattern.  Starts at 0.
-			unsigned int row;
+			unsigned long row;
 
 			/// True if the end of the song has been reached.
 			bool end;
